requirements.txt
flask
boto3
aws-cdk-lib
constructs
pyyaml
jsii